restructuring and renaming in preparation for travis-ci + sonar
