+++ .coveragerc
@@ -1,0 +1,13 @@
+[run]
+branch = True
+omit =
+    /test/*
+    */migrations/*
+    *__init__.py*
+    *__main__.py*
+
+[report]
+exclude_lines =
+    # Ignore imports
+    from
+    import
