+++ .travis.yml
@@ -1,13 +1,11 @@
-# ref: https://docs.travis-ci.com/user/languages/python
 language: python
 python:
-  - "3.2"
-  - "3.3"
-  - "3.4"
-  - "3.5"
-  #- "3.5-dev" # 3.5 development branch
-  #- "nightly" # points to the latest development branch e.g. 3.6-dev
-# command to install dependencies
-install: "pip install -r requirements.txt"
-# command to run tests
-script: nosetests
+  - "3.8"
+
+install:
+  - pip install -r requirements.txt
+  - pip install -r test-requirements.txt
+script:
+  # other script steps might be done before running the actual analysis
+  - pytest --cov=server && coverage xml
+  - sonar-scanner
