+++ server/models/layer.py
@@ -5,9 +5,9 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server.models.layer_entry import LayerEntry  # noqa: F401,E501
-from swagger_server import util
+from server.models.base_model_ import Model
+from server.models.layer_entry import LayerEntry  # noqa: F401,E501
+from server import util
 
 
 class Layer(Model):
