+++ setup.py
@@ -3,7 +3,7 @@
 import sys
 from setuptools import setup, find_packages
 
-NAME = "swagger_server"
+NAME = "server"
 VERSION = "1.0.0"
 # To install the library, run the following
 #
@@ -26,7 +26,7 @@ setup(
     package_data={'': ['swagger/swagger.yaml']},
     include_package_data=True,
     entry_points={
-        'console_scripts': ['swagger_server=swagger_server.__main__:main']},
+        'console_scripts': ['server=server.__main__:main']},
     long_description="""\
     REST API for Advanced Software Engineering FS21 Project
     """
