+++ sonar-project.properties
@@ -1,0 +1,17 @@
+# must be unique in a given SonarQube instance
+sonar.projectKey=cdeiac_sonarqube-sample
+sonar.login=$SONAR_TOKEN
+
+# links
+sonar.links.ci=https://www.travis-ci.com/github/cdeiac/sonarqube-sample
+sonar.links.issue=https://github.com/cdeiac/sonarqube-sample/issues
+sonar.links.scm=https://github.com/cdeiac/sonarqube-sample/tree/main
+
+# project settings
+sonar.sources=./server
+sonar.tests=./tests
+sonar.language=py
+sonar.exclusions=.swagger-codegen/**,setup.py
+
+# testing
+sonar.python.coverage.reportPaths=coverage.xml
