--- swagger_server/models/__init__.py
@@ -1,10 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-from __future__ import absolute_import
-# import models into model package
-from swagger_server.models.host import Host
-from swagger_server.models.layer import Layer
-from swagger_server.models.layer_entry import LayerEntry
-from swagger_server.models.listing import Listing
-from swagger_server.models.review import Review
