+++ test-requirements.txt
@@ -1,6 +1,8 @@
 flask_testing==0.8.0
 coverage>=4.0.3
 nose>=1.3.7
+pytest>=6.2.2
+pytest-cov>=2.11.1
 pluggy>=0.3.1
 py>=1.4.31
 randomize>=0.13
