+++ test/__init__.py
@@ -3,14 +3,14 @@ import logging
 import connexion
 from flask_testing import TestCase
 
-from swagger_server.encoder import JSONEncoder
+from server.encoder import JSONEncoder
 
 
 class BaseTestCase(TestCase):
 
     def create_app(self):
         logging.getLogger('connexion.operation').setLevel('ERROR')
-        app = connexion.App(__name__, specification_dir='../swagger/')
+        app = connexion.App(__name__, specification_dir='../server/swagger/')
         app.app.json_encoder = JSONEncoder
         app.add_api('swagger.yaml', pythonic_params=True)
         return app.app
