+++ test/test_hosts_controller.py
@@ -2,11 +2,7 @@
 
 from __future__ import absolute_import
 
-from flask import json
-from six import BytesIO
-
-from swagger_server.models.host import Host  # noqa: E501
-from swagger_server.test import BaseTestCase
+from test import BaseTestCase
 
 
 class TestHostsController(BaseTestCase):
