+++ test/test_recommendations_controller.py
@@ -2,11 +2,7 @@
 
 from __future__ import absolute_import
 
-from flask import json
-from six import BytesIO
-
-from swagger_server.models.listing import Listing  # noqa: E501
-from swagger_server.test import BaseTestCase
+from test import BaseTestCase
 
 
 class TestRecommendationsController(BaseTestCase):
