+++ tox.ini
@@ -1,10 +1,11 @@
 [tox]
-envlist = py35
+envlist = python3.8
 
 [testenv]
 deps=-r{toxinidir}/requirements.txt
      -r{toxinidir}/test-requirements.txt
 
-commands=
-   nosetests \
-      []
+[pytest]
+addopts = -p no:warnings
+
+commands = pytest
